# Performance and other fixes - part 1