--- a/sabnzbd/__init__.py
+++ b/sabnzbd/__init__.py
@@ -146,7 +146,7 @@ WEBUI_READY = False
 LAST_WARNING = None
 LAST_ERROR = None
 EXTERNAL_IPV6 = False
-LAST_HISTORY_CALL = time.time()
+LAST_HISTORY_UPDATE = time.time()
 
 __INITIALIZED__ = False
 __SHUTTING_DOWN__ = False