--- a/sabnzbd/api.py
+++ b/sabnzbd/api.py
@@ -474,10 +474,10 @@ def _api_history(name, output, kwargs):
     search = kwargs.get('search')
     failed_only = kwargs.get('failed_only')
     categories = kwargs.get('category')
-    last_history_call = kwargs.get('last_history_call', 0)
+    last_history_update = kwargs.get('last_history_update', 0)
 
     # Do we need to send anything?
-    if(int(last_history_call) == int(sabnzbd.LAST_HISTORY_CALL)):
+    if(int(last_history_update) == int(sabnzbd.LAST_HISTORY_UPDATE)):
         return report(output, keyword='history', data=False)
 
     if categories and not isinstance(categories, list):
@@ -498,14 +498,14 @@ def _api_history(name, output, kwargs):
             if special in ('all', 'completed'):
                 history_db.remove_completed(search)
             # Update the last check time
-            sabnzbd.LAST_HISTORY_CALL = time.time()
+            sabnzbd.LAST_HISTORY_UPDATE = time.time()
             return report(output)
         elif value:
             jobs = value.split(',')
             for job in jobs:
                 del_hist_job(job, del_files)
             # Update the last check time
-            sabnzbd.LAST_HISTORY_CALL = time.time()
+            sabnzbd.LAST_HISTORY_UPDATE = time.time()
             return report(output)
         else:
             return report(output, _MSG_NO_VALUE)
@@ -521,7 +521,7 @@ def _api_history(name, output, kwargs):
                                                                               search=search, failed_only=failed_only,
                                                                               categories=categories,
                                                                               output=output)
-        history['last_history_call'] = int(sabnzbd.LAST_HISTORY_CALL)
+        history['last_history_update'] = int(sabnzbd.LAST_HISTORY_UPDATE)
         return report(output, keyword='history', data=remove_callable(history))
     else:
         return report(output, _MSG_NOT_IMPLEMENTED)