--- a/sabnzbd/nzbstuff.py
+++ b/sabnzbd/nzbstuff.py
@@ -1504,7 +1504,7 @@ class NzbObject(TryList):
 
     def set_action_line(self, action=None, msg=None):
         # Update the last check time
-        sabnzbd.LAST_HISTORY_CALL = time.time()
+        sabnzbd.LAST_HISTORY_UPDATE = time.time()
         if action and msg:
             self.action_line = '%s: %s' % (action, msg)
         else: