Without changing too much of the actual downloader code yet, there is performance to gain!

- Remove ```asserts``` throughout the code. Even though the ```-OO``` option is specified in the file header, it is not actually applied in many cases. For example JCFP's Ubuntu package will not use it. Gives performance boost and reduction of CPU usage of 10-20% in these cases.
- Decoder had a small queue of only max. 10 jobs, but on a fast connection with 50 connections this will fill up instantly and the decoder will pause the downloader because the Queue is full. Now the decoder queue is part of the article-cache so it is basically not limited. Speed before was limited to 25MB/s, now I get 45MB/s (although NZBGet still gets 75MB/s on the same NZB).
- The downloader-while-loop was without a "speedlimit", causing it too loop too often without doing any action but still performing all the checks. Especially on slow connections it makes no sense to check so often. By limiting the while loop on my 60Mbit connection the CPU usage dropped from 31% to 19%.
- Increasing the receive buffer size allows for making the downloader-while-loop even slower. But for some reason this is not possible for SSL connections, due to some kind of lock. 

The current speedlimit on the downloader-while-loop does make SSL connections slower on high speed (above 25MB/s). 
For the next iteration going to try other approaches to make the downloader-while-loop less heavy or maybe even dependent on the actual maximum speed.