@Safihre 2016-05-02
Also implemented:

- Only 1 error message will be shown when there's a connection problem, not 1 per connection.
- Removed the support for UUEncode since it would only be interesting if we actually supported multipart-UUencode, but we don't.. Only the (never seen) singlepart one. Since we only got 1 complaint about multioar-UUencode, clearly nobody is using that either.
- Added a 'Restore Defaults' button to the Switches page