@sanderjo 2016-05-02
@Safihre 

I got your code (https://github.com/Safihre/sabnzbd.git) and ran it on an old Atom. A few different newsservers, no SSL, Fast Ethernet on a 100/100 Mbps Fiber connection. 

Resulting speed with your code: 
`Downloaded in 2 mins 59 seconds at an average of 5.5 MB/s`

I switched back to SABnzbd 1.0.1RC1, and did the same download: 
`Downloaded in 2 mins 46 seconds at an average of 6.0 MB/s`

So your code seems to be slower on my setup?