@Safihre 2016-05-02
@sanderjo Damm, did not expect that. Seems I have to go back to the drawing board. 
I removed that last commit about the downloader. Could you maybe try again and compare 1.0.1 vs. mine? Might have to remove it and clone it again since probably it doesn't like that I removed a commit.
Just to make sure: on both 1.0.1 and when you run my version yEnc is working (shows a checkmark on first Config page).
This NZB is also undamaged so no fetching of extra blocks, right?

I don't have a slower system myself to test on, so indeed will need input from you :+1: