@thezoggy 2016-05-02
wonder if the atom box isnt cpu bound but io bound. thus the optimizations arent really going to do much for it, if thats the case. move d/l directory to flash/ssd/etc? check stats.. cpu maxxed? memory maxxed? io maxxed?

@Safihre just force push to update your repo. to slow down your box you can do all sorts of stuff.. but an easy one is to just limit process affinity to one cpu.. so that way you have more of a chance to be cpu starved.