@shypike 2016-05-03
@sanderjo Did you set a cache and how big is it?
Without a cache there might not be a sufficient decoder queue.