@sanderjo 2016-05-03
git pull (correct?):
```
sander@netbook:~/git/safihre-sab$ git pull
From https://github.com/Safihre/sabnzbd
 + ec50390...16d9516 develop    -> origin/develop  (forced update)
Already up-to-date.

```
Result:
`Downloaded in 2 mins 51 seconds at an average of 5.8 MB/s
`

CPU load during download doesn't look bad; see screenshot:

![screenshot from 2016-05-03 08-07-10](https://cloud.githubusercontent.com/assets/1273502/14975886/b6fd1a78-1106-11e6-8471-248dd47e1cce.png)