@sanderjo 2016-05-03
@shypike 

Yes, cache was set during the tests: 777M
The Atom system has 1GB RAM. I will also try a lower cache. I don't expect anything from that, because there is still 792 MB RAM free.


```
sander@netbook:~$ free -m
             total       used       free     shared    buffers     cached
Mem:           991        709        282         36        140        369
-/+ buffers/cache:        198        792
Swap:         2212         45       2166

```