@Safihre 2016-05-03
@shypike I changed the code for the decoder queue since yesterday morning, now when no cache is set it will use the old maximum decoder length.

@sanderjo Maybe you can try again tonight with both 1.0.1 and mine right after the other? To make sure that there's no other things at play. Interesting that load is not maximum, indeed it might be more I/O bound than CPU bound? I would expect it to consume all CPU power for decoding/assembling, but it doesn't?