@sanderjo 2016-05-03
Hmmm: I downloaded with a certain other NZB-downloader, and I got 7.5 MB/s. So: better speed, but no line speed.