@Safihre 2016-05-05
Added the ```asserts``` back using the ```if 0:```. 
Also found that using ```sys.getsizeof``` did not return the actual size of the data-list that was being put in the decoder/articlecache, most probably because it was just a reference instead of the real object.
Found that multiplying it by 32 is a small overestimation (about 2-3%), but better to be on safe side.
Tested this with NZB's with all kind of article sizes.

Will continue to work on the real performance improvements, so far it's mostly fixes. Updated the title accordingly.