@Safihre 2016-05-05
@sanderjo I made my VM to only use 20% of CPU to reproduce a slow system and the changes that are still in the request do improve the speed on my system.
```
#1 (small)
Downloaded in 53 seconds at an average of 7.0 MB/s
vs
Downloaded in 1 min 1 sec at an average of 5.7 MB/s

#2 (bigger)
Downloaded in 2 mins 20 seconds at an average of 7.1 MB/s
vs
Downloaded in 3 mins 2 seconds at an average of 5.6 MB/s
```
That's probably just from increasing the decoder size. Cache was set to 150M.