@thezoggy 2016-05-10
i should check out the fixes on this laptop, it is constantly pegged at 100% trying to download which ends up limiting me to 12-14mb/s... you can easily tell when your not getting max speed as the download bar is super spikey..  when downloading in sab I see windows defender takes up like 24% cpu... wtf..