@Safihre 2016-05-11
Yep same here, another reason not to use defender. In other anti virus you can at least turn of the checking if nntp connections combined with excluding the temp. Folder..