@thezoggy 2016-05-11
i disabled defender via regedit (since i couldnt do global policy on my version of win10).
but we should check if this provides a 'fix' and recommend people adding an exclusion for the incomplete folder / sab exe etc... maybe offer a button to auto set it?
http://windows.microsoft.com/en-us/windows-10/add-an-exclusion-to-windows-defender